+++ src/components/barista-components/notes.txt
