+++ src/components/main-view-components/barista-view.jsx
@@ -1,11 +1,13 @@
 import {
     Box,
   } from "@mui/material";
+import { Route, Roputes, useNavigate } from "react-router-dom";
 import CompositeTitleHeader from "../shared-components/composite-header";
 import StickyFooter from "../shared-components/sticky-footer";
 import "./../../App.css";
 import "./general.css";
 import HoursAdder from "../shared-components/hours-adder";
+import CheckIngredients from "../barista-components/check-ingredients";
 export default function BaristaView() {
     return (
         <div className="App">
@@ -13,7 +15,9 @@ export default function BaristaView() {
                 <CompositeTitleHeader />
             </Box>
             <Box className="WorkingViewContainer">
-                Insert Custom Components Here. Can change styling of WorkingViewContainer if Needed
+                <Routes>
+                    <Route path="check-ingredients" element={<CheckIngredients className="DrinkAdderContainer" ingredientClassName="NewDrinkIngridentScroller"/>} />
+                </Routes>
             </Box>
             <Box className="StickyFooterContainer">
                 <StickyFooter>
